Move Preferences into Config namespace
